Music Update

+++ src/main/java/com/beyondbell/bugisoft/eventHandling/GuildMessageReceivedEventHandler.kt
@@ -2,6 +2,7 @@ package com.beyondbell.bugisoft.eventHandling
 
 import com.beyondbell.bugisoft.DEFAULT_INVITE_TIME
 import com.beyondbell.bugisoft.invite.createInvite
+import com.beyondbell.bugisoft.music.Music
 import com.beyondbell.bugisoft.ping.Ping
 import com.beyondbell.bugisoft.utilities.getParameters
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent
@@ -18,7 +19,7 @@ fun handleGuildMessageReceivedEvent(event: GuildMessageReceivedEvent) {
 	}
 
 	// Checks for Prefixes
-	if (event.message.contentRaw[0] != '!') {
+	if (event.message.contentRaw[0] != '!' && event.message.contentRaw[0] != ';') {
 		return
 	}
 
@@ -28,6 +29,7 @@ fun handleGuildMessageReceivedEvent(event: GuildMessageReceivedEvent) {
 		"!" -> {
 			when (parameters[1].toLowerCase()) {
 				"invite" -> {
+					event.message.delete().queue()
 					if (parameters.size == 3) {
 						try {
 							createInvite(event, Integer.parseInt(parameters[2]))
@@ -37,11 +39,48 @@ fun handleGuildMessageReceivedEvent(event: GuildMessageReceivedEvent) {
 					} else {
 						createInvite(event, DEFAULT_INVITE_TIME)
 					}
-					event.message.delete().queue()
 				}
 				"ping" -> {
-					Ping.ping(event)
 					event.message.delete().queue()
+					Ping.ping(event)
+				}
+			}
+		}
+		";" -> {
+			when (parameters[1].toLowerCase()) {
+				"join" -> {
+					event.message.delete().queue()
+					Music.join(event)
+				}
+				"leave" -> {
+					event.message.delete().queue()
+					Music.leave(event)
+				}
+				"play" -> {
+					event.message.delete().queue()
+					if (parameters.size > 2) {
+						Music.play(event, parameters.copyOfRange(2, parameters.size))
+					}
+				}
+				"clear" -> {
+					event.message.delete().queue()
+					Music.clear(event.guild.idLong)
+				}
+				"skip" -> {
+					event.message.delete().queue()
+					if (parameters.size > 2) {
+						Music.skip(event.guild.idLong, Integer.valueOf(parameters[2]))
+					} else {
+						Music.skip(event.guild.idLong, 1)
+					}
+				}
+				"shuffle" -> {
+					event.message.delete().queue()
+					Music.shuffle(event.guild.idLong)
+				}
+				"repeat" -> {
+					event.message.delete().queue()
+					Music.toggleRepeatMode(event.guild.idLong)
 				}
 			}
 		}
